wip: add tests, bugfixes
