+++ todo.md
@@ -1,0 +1,8 @@
+# TODO
+
+## `dump-documents`
+
+### `archive` option
+
+- [ ] configure archive compression format via config file
+- [ ] configure archive compression format via command line option
